Optimize key types in documents

+++ io/shard/document_info.go
@@ -3,5 +3,5 @@ package shard
 type DocumentInfo struct {
 	Start int64
 	Pos   int64
-	Len   uint32
+	Len   uint16
 }
